requirements.txt
tortoise-orm[asyncpg]==0.19.2
pylint==2.15.10
yapf==0.32.0
temporalio==1.0.0
click==8.1.3
aiohttp==3.8.3
openapi-python-client==0.13.1
dataclass-wizard==0.22.2
fastapi==0.89.1
fastapi-pagination==0.11.2
Jinja2==3.1.2
hypercorn==0.14.3
setuptools==58.2.0
pydantic==1.10.4
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
itsdangerous==2.1.2
PyYAML==6.0
beautifulsoup4==4.11.2